+++ src/styles/Gallery/Gallery.ts
