+++ src/styles/Gallery/GalleryItem.ts
