+++ src/styles/Gallery/GalleryList.ts
