+++ src/styles/Gallery/GalleryModal.ts
